+++ src/bda/plone/checkout/browser/form.py
@@ -10,6 +10,7 @@ from yafowil.base import factory
 from yafowil.plone.form import Form
 from yafowil.yaml import parse_from_YAML
 from bda.plone.cart import readcookie
+from bda.plone.cart import get_data_provider
 from bda.plone.checkout import CheckoutDone
 from bda.plone.checkout import message_factory as _
 from bda.plone.checkout.interfaces import CheckoutError
@@ -61,6 +62,7 @@ class FieldsProvider(FormContext):
     fields_name = ''
     message_factory = _
     ignore_on_save = False
+    skip = False
 
     def __init__(self, context, request):
         self.context = context
@@ -88,6 +90,8 @@ class FieldsProvider(FormContext):
         return ret
 
     def extend(self, form):
+        if self.skip:
+            return
         fields = parse_from_YAML(self.fields_template,
                                  self, self.message_factory)
         form[self.fields_name] = fields
@@ -214,6 +218,11 @@ provider_registry.add(ShippingSelection)
 class PaymentSelection(FieldsProvider):
     fields_template = 'bda.plone.checkout.browser:forms/payment_selection.yaml'
     fields_name = 'payment_selection'
+
+    @property
+    def skip(self):
+        cart_data = get_data_provider(self.context, self.request)
+        return not cart_data.total
 
     @property
     def payments(self):